refactor(next): write cli in ts
