+++ packages/react-server-next/tsconfig.json
@@ -2,7 +2,12 @@
   "extends": "../../tsconfig.base.json",
   "include": ["src", "vitest.config.ts", "tsup.config.ts"],
   "compilerOptions": {
-    "types": ["vite/client", "react/experimental", "react-dom/experimental"],
+    "types": [
+      "vite/client",
+      "react/experimental",
+      "react-dom/experimental",
+      "@types/node"
+    ],
     "jsx": "react-jsx"
   }
 }
